refactor: userRepository dto 의존성 제거
