+++ app/src/main/java/com/smilehunter/ablebody/domain/repository/ManageRepository.kt
@@ -1,7 +1,7 @@
 package com.smilehunter.ablebody.domain.repository
 
-import com.smilehunter.ablebody.network.model.request.ReportRequest
 import com.smilehunter.ablebody.network.model.ReportResponse
+import com.smilehunter.ablebody.network.model.request.ReportRequest
 
 interface ManageRepository {
     suspend fun report(reportRequest: ReportRequest): ReportResponse
