+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/AddCouponUseCase.kt
@@ -1,6 +1,5 @@
 package com.smilehunter.ablebody.domain.usecase
 
-import com.smilehunter.ablebody.network.model.AbleBodyResponse
 import com.smilehunter.ablebody.domain.repository.UserRepository
 import com.smilehunter.ablebody.network.di.AbleBodyDispatcher
 import com.smilehunter.ablebody.network.di.Dispatcher
@@ -14,24 +13,6 @@ class AddCouponUseCase @Inject constructor(
 ) {
 
     suspend operator fun invoke(couponCode: String): CouponStatus = withContext(ioDispatcher) {
-        val response = userRepository.addCouponByCouponCode(couponCode)
-        if (response.success) {
-            return@withContext CouponStatus.SUCCESS
-        }
-
-        when(response.errorCode) {
-            AbleBodyResponse.ErrorCode.UNABLE_COUPON -> CouponStatus.UNABLE_COUPON
-            AbleBodyResponse.ErrorCode.INVALID_COUPON_CODE -> CouponStatus.INVALID_COUPON_CODE
-            AbleBodyResponse.ErrorCode.ALREADY_EXIST_COUPON -> CouponStatus.ALREADY_EXIST_COUPON
-            else -> CouponStatus.UNKNOWN_ERROR
-        }
+        userRepository.addCouponByCouponCode(couponCode)
     }
-}
-
-enum class CouponStatus {
-    SUCCESS,
-    UNKNOWN_ERROR,
-    INVALID_COUPON_CODE,
-    ALREADY_EXIST_COUPON,
-    UNABLE_COUPON,
 }
