+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/CodyItemPagerUseCase.kt
@@ -7,15 +7,15 @@ import androidx.paging.PagingSource
 import androidx.paging.PagingState
 import com.smilehunter.ablebody.data.model.Gender
 import com.smilehunter.ablebody.data.model.HomeCategory
-import com.smilehunter.ablebody.network.model.response.BrandDetailCodyResponseData
-import com.smilehunter.ablebody.network.model.response.FindCodyResponseData
-import com.smilehunter.ablebody.network.model.response.ReadBookmarkCodyData
-import com.smilehunter.ablebody.network.model.response.SearchCodyResponseData
+import com.smilehunter.ablebody.domain.model.CodyItemData
 import com.smilehunter.ablebody.domain.repository.BookmarkRepository
 import com.smilehunter.ablebody.domain.repository.BrandRepository
 import com.smilehunter.ablebody.domain.repository.FindCodyRepository
 import com.smilehunter.ablebody.domain.repository.SearchRepository
-import com.smilehunter.ablebody.domain.model.CodyItemData
+import com.smilehunter.ablebody.network.model.response.BrandDetailCodyResponseData
+import com.smilehunter.ablebody.network.model.response.FindCodyResponseData
+import com.smilehunter.ablebody.network.model.response.ReadBookmarkCodyData
+import com.smilehunter.ablebody.network.model.response.SearchCodyResponseData
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.withContext
