+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/EditProfileResult.kt
@@ -1,0 +1,5 @@
+package com.smilehunter.ablebody.domain.usecase
+
+enum class EditProfileResult {
+    SUCCESS, FAIL
+}
