+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/EditProfileUseCase.kt
@@ -1,9 +1,9 @@
 package com.smilehunter.ablebody.domain.usecase
 
-import com.smilehunter.ablebody.network.model.request.EditProfile
 import com.smilehunter.ablebody.domain.repository.UserRepository
 import com.smilehunter.ablebody.network.di.AbleBodyDispatcher.IO
 import com.smilehunter.ablebody.network.di.Dispatcher
+import com.smilehunter.ablebody.network.model.request.EditProfile
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import java.io.InputStream
@@ -17,22 +17,18 @@ class EditProfileUseCase @Inject constructor(
     suspend operator fun invoke(
         editProfile: EditProfile,
         profileImageInputStream: InputStream?
-    ): Status = withContext(ioDispatcher) {
-        val response = userRepository.editProfile(
+    ): EditProfileResult = withContext(ioDispatcher) {
+        val result = userRepository.editProfile(
             editProfile = editProfile,
             profileImageInputStream = profileImageInputStream
         )
         profileImageInputStream?.close()
 
-        if (response.success) {
-            return@withContext Status.SUCCESS
-        }
 
-        Status.FAIL
+        if (result) {
+            return@withContext EditProfileResult.SUCCESS
+        }
+        EditProfileResult.FAIL
     }
 
-}
-
-enum class Status {
-    SUCCESS, FAIL
 }
