+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/GetCommentListUseCase.kt
@@ -1,7 +1,7 @@
 package com.smilehunter.ablebody.domain.usecase
 
-import com.smilehunter.ablebody.domain.repository.CreatorDetailRepository
 import com.smilehunter.ablebody.domain.model.CommentListData
+import com.smilehunter.ablebody.domain.repository.CreatorDetailRepository
 import com.smilehunter.ablebody.network.model.response.CreatorDetailResponseData
 import javax.inject.Inject
