+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/GetCreatorDetailDataListUseCase.kt
@@ -1,8 +1,8 @@
 package com.smilehunter.ablebody.domain.usecase
 
 import com.smilehunter.ablebody.data.model.ItemChildCategory
-import com.smilehunter.ablebody.domain.repository.CreatorDetailRepository
 import com.smilehunter.ablebody.domain.model.CreatorDetailData
+import com.smilehunter.ablebody.domain.repository.CreatorDetailRepository
 import com.smilehunter.ablebody.network.model.response.CreatorDetailResponseData
 import javax.inject.Inject
