+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/GetLikeListUseCase.kt
@@ -1,8 +1,8 @@
 package com.smilehunter.ablebody.domain.usecase
 
-import com.smilehunter.ablebody.domain.repository.LikeListRepository
 import com.smilehunter.ablebody.domain.model.LikeListData
 import com.smilehunter.ablebody.domain.model.LikedLocations
+import com.smilehunter.ablebody.domain.repository.LikeListRepository
 import com.smilehunter.ablebody.network.model.response.CreatorDetailLikeUsersResponseData
 import javax.inject.Inject
