+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/GetMyDeliveryAddressUseCase.kt
@@ -1,10 +1,9 @@
 package com.smilehunter.ablebody.domain.usecase
 
-import com.smilehunter.ablebody.domain.repository.UserRepository
 import com.smilehunter.ablebody.domain.model.DeliveryAddressData
+import com.smilehunter.ablebody.domain.repository.UserRepository
 import com.smilehunter.ablebody.network.di.AbleBodyDispatcher
 import com.smilehunter.ablebody.network.di.Dispatcher
-import com.smilehunter.ablebody.network.model.response.GetAddressResponseData
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
@@ -15,17 +14,6 @@ class GetMyDeliveryAddressUseCase @Inject constructor(
 ) {
 
     suspend operator fun invoke(): DeliveryAddressData = withContext(ioDispatcher) {
-        userRepository.getMyAddress().data!!.toDomain()
+        userRepository.getMyAddress()
     }
 }
-
-private fun GetAddressResponseData.toDomain() =
-    DeliveryAddressData(
-        id = id,
-        userName = receiverName,
-        roadAddress = addressInfo,
-        roadDetailAddress = detailAddress,
-        zipCode = zipCode,
-        phoneNumber = phoneNum,
-        deliveryRequestMessage = deliveryRequest
-    )
