+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/GetReceiptUseCase.kt
@@ -1,7 +1,7 @@
 package com.smilehunter.ablebody.domain.usecase
 
-import com.smilehunter.ablebody.domain.repository.OrderManagementRepository
 import com.smilehunter.ablebody.domain.model.ReceiptData
+import com.smilehunter.ablebody.domain.repository.OrderManagementRepository
 import com.smilehunter.ablebody.network.di.AbleBodyDispatcher
 import com.smilehunter.ablebody.network.di.Dispatcher
 import com.smilehunter.ablebody.network.model.response.GetOrderListDetailResponseData
