+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/GetSearchHistoryUseCase.kt
@@ -1,7 +1,7 @@
 package com.smilehunter.ablebody.domain.usecase
 
-import com.smilehunter.ablebody.domain.repository.SearchRepository
 import com.smilehunter.ablebody.domain.model.SearchHistoryQuery
+import com.smilehunter.ablebody.domain.repository.SearchRepository
 import com.smilehunter.ablebody.network.di.AbleBodyDispatcher
 import com.smilehunter.ablebody.network.di.Dispatcher
 import kotlinx.coroutines.CoroutineDispatcher
