+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/GetUserBoardPagerUseCase.kt
@@ -5,8 +5,8 @@ import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
-import com.smilehunter.ablebody.domain.repository.UserRepository
 import com.smilehunter.ablebody.domain.model.UserBoardData
+import com.smilehunter.ablebody.domain.repository.UserRepository
 import com.smilehunter.ablebody.network.di.AbleBodyDispatcher
 import com.smilehunter.ablebody.network.di.Dispatcher
 import com.smilehunter.ablebody.network.model.response.GetMyBoardResponseData
