+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/NotificationItemPagerUseCase.kt
@@ -5,8 +5,8 @@ import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
-import com.smilehunter.ablebody.domain.repository.NotificationRepository
 import com.smilehunter.ablebody.domain.model.NotificationItemData
+import com.smilehunter.ablebody.domain.repository.NotificationRepository
 import com.smilehunter.ablebody.network.model.response.GetMyNotiResponseData
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
