+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/ProductItemPagerUseCase.kt
@@ -9,17 +9,17 @@ import com.smilehunter.ablebody.data.model.ItemChildCategory
 import com.smilehunter.ablebody.data.model.ItemGender
 import com.smilehunter.ablebody.data.model.ItemParentCategory
 import com.smilehunter.ablebody.data.model.SortingMethod
-import com.smilehunter.ablebody.network.model.response.BrandDetailItemResponseData
-import com.smilehunter.ablebody.network.model.response.FindItemResponseData
-import com.smilehunter.ablebody.network.model.response.ReadBookmarkItemData
-import com.smilehunter.ablebody.network.model.response.SearchItemResponseData
+import com.smilehunter.ablebody.domain.model.ProductItemData
 import com.smilehunter.ablebody.domain.repository.BookmarkRepository
 import com.smilehunter.ablebody.domain.repository.BrandRepository
 import com.smilehunter.ablebody.domain.repository.FindItemRepository
 import com.smilehunter.ablebody.domain.repository.SearchRepository
-import com.smilehunter.ablebody.domain.model.ProductItemData
 import com.smilehunter.ablebody.network.di.AbleBodyDispatcher
 import com.smilehunter.ablebody.network.di.Dispatcher
+import com.smilehunter.ablebody.network.model.response.BrandDetailItemResponseData
+import com.smilehunter.ablebody.network.model.response.FindItemResponseData
+import com.smilehunter.ablebody.network.model.response.ReadBookmarkItemData
+import com.smilehunter.ablebody.network.model.response.SearchItemResponseData
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.withContext
